requirements.txt
transformers==4.52.4
torch==2.6
openai==1.82.1
python-dotenv==1.1.0
datasets==3.6.0
huggingface-hub==0.32.3
tiktoken==0.9.0
chromadb==1.0.12
sentence-transformers==4.1.0
gradio==5.32.1
weaviate-client[embedded]